+++ mcmd/__main__.py
@@ -14,7 +14,7 @@ import logging
 import signal
 
 from mcmd.io import io
-from mcmd.argparser import parse_args
+from mcmd.core.argparser import parse_args
 from mcmd.config.loader import load_config
 from mcmd.io.io import set_debug
 from mcmd.io.logging import set_level
