+++ mcmd/commands/disable.py
@@ -1,7 +1,7 @@
 from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight
 from mcmd.molgenis.resources import ensure_resource_exists, ResourceType
