+++ mcmd/commands/history.py
@@ -1,6 +1,6 @@
-from mcmd import history as hist
+from mcmd.core import history as hist
 from mcmd.io import io
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.logging import get_logger
