+++ mcmd/commands/make.py
@@ -1,10 +1,10 @@
 from mcmd.io import io
 from mcmd.molgenis import api
 from mcmd.molgenis.client import post, get
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 from mcmd.utils.utils import lower_kebab, upper_snake
