+++ mcmd/commands/ping.py
@@ -1,10 +1,10 @@
 from colorama import Fore
 
 import mcmd.config.config as config
-from mcmd.command import command
+from mcmd.core.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io.io import highlight
-from mcmd.utils.errors import McmdError
+from mcmd.core.errors import McmdError
 from mcmd.version import molgenis_version
