+++ mcmd/core/errors.py
