+++ mcmd/molgenis/version.py
@@ -6,7 +6,7 @@ import requests
 from requests import HTTPError
 
 from mcmd.config import config
-from mcmd.utils.errors import McmdError, MolgenisOfflineError
+from mcmd.core.errors import McmdError, MolgenisOfflineError
 
 _version_number = None
 _version = None
